backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.10.3,<3.0.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1

# AI & NLP
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.16,<0.2.0
langchain-community>=0.0.14,<0.1.0
google-generativeai>=0.3.1,<0.4.0

# Data Processing
networkx==3.2.1
pandas==2.2.0
numpy==1.26.3

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# CORS
fastapi-cors==0.0.6
